refactor(core): reorganize files for better structure and clarity
