+++ core/src/main/kotlin/si/inova/tws/core/data/view/WebContent.kt
@@ -14,7 +14,7 @@
  *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package si.inova.tws.core.data
+package si.inova.tws.core.data.view
 
 sealed class WebContent {
     data class Url(
