+++ core/src/main/kotlin/si/inova/tws/core/data/view/WebViewNavigator.kt
@@ -14,7 +14,7 @@
  *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package si.inova.tws.core.data
+package si.inova.tws.core.data.view
 
 import android.webkit.WebView
 import androidx.compose.runtime.Composable
